+++ packages/ko/src/core/config.ts
@@ -65,7 +65,6 @@ class Config {
         disableLazyImports: false,
       },
       autoPolyfills: false,
-      logLevel: 'error',
     };
   }
 }