+++ packages/ko/src/webpack/plugins.ts
@@ -10,6 +10,7 @@ import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 import AutoPolyfillsWebpackPlugin from 'auto-polyfills-webpack-plugin';
 import FriendlyErrorsWebpackPlugin from '@nuxt/friendly-errors-webpack-plugin';
 import PurgeCacheWebpackPlugin from './custom-plugins/purge-cache-webpack-plugin';
+import ErrorOverlayPlugin from 'error-overlay-webpack-plugin';
 import { IWebpackOptions } from '../types';
 
 function getPlugins(opts: IWebpackOptions) {
@@ -95,6 +96,7 @@ function getPlugins(opts: IWebpackOptions) {
         },
       }),
     !isProd && new PurgeCacheWebpackPlugin(),
+    !isProd && new ErrorOverlayPlugin(),
     isProd &&
       autoPolyfills &&
       (typeof autoPolyfills === 'boolean'